+++ mhs-reference-implementation/mhs/routing/sds_exception.py
@@ -1,6 +1,6 @@
 
 
-class RoutingException(Exception):
+class SDSException(Exception):
     """
     A RoutingException is thrown when an issue arises with the LDAP response,
     particularly when no response is returned from the LDAP query.
