+++ mhs-reference-implementation/mhs/routing/tests/test_routing_reliability.py
@@ -2,9 +2,9 @@ import unittest
 
 from utilities.test_utilities import async_test
 from mhs.routing import sds_handler
-import mhs.routing.routing_reliabillity as rar
+import mhs.routing.routing_reliability as rar
 import mhs.routing.tests.ldap_mocks as mocks
-import mhs.routing.routing_exception as route_exception
+import mhs.routing.sds_exception as route_exception
 import mhs.routing.dictionary_cache as dict_cache
 
 PARTY_KEY = " AP4RTY-K33Y"
@@ -41,12 +41,12 @@ class TestRoutingReliability(unittest.TestCase):
 
     @async_test
     async def test_get_routing_bad_ods_code(self):
-        with self.assertRaises(route_exception.RoutingException):
+        with self.assertRaises(route_exception.SDSException):
             await self.router.get_end_point("bad code", INTERACTION_ID)
 
     @async_test
     async def test_get_routing_bad_interaction_id(self):
-        with self.assertRaises(route_exception.RoutingException):
+        with self.assertRaises(route_exception.SDSException):
             await self.router.get_end_point(ODS_CODE, "bad interaction")
 
     @async_test
@@ -56,12 +56,12 @@ class TestRoutingReliability(unittest.TestCase):
 
     @async_test
     async def test_get_reliability_bad_ods_code(self):
-        with self.assertRaises(route_exception.RoutingException):
+        with self.assertRaises(route_exception.SDSException):
             await self.router.get_reliability("bad code", INTERACTION_ID)
 
     @async_test
     async def test_get_reliability_bad_interaction_id(self):
-        with self.assertRaises(route_exception.RoutingException):
+        with self.assertRaises(route_exception.SDSException):
             await self.router.get_reliability(ODS_CODE, "whew")
 
     @async_test
